backend/requirements.txt
flask
flask-cors
mysql-connector-python
flask-bcrypt
dnspython
python-dotenv
gunicorn
PyJWT
marshmallow